Allow specifying a channel key for access to a channel
There are some cases where an irc channel may have a key associated with it.
This key is required for access to the channel. If it is specified in config,
the key will be passed when the channel join is attempted. If no key is
present, none will be passed. No key is the default behavior.

Bug: T122104
Change-Id: I14458b9df9d7a15dc83a397e6b4ce7609f3b449d